requirements.txt
aiokafka==0.12.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncio==3.4.3
bcrypt==3.2.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.8
filelock==3.17.0
greenlet==3.1.1
groq==0.16.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
identify==2.6.6
idna==3.10
kafka-python==2.0.2
nodeenv==1.9.1
packaging==24.2
passlib==1.7.4
platformdirs==4.3.6
pre_commit==4.1.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.1
watchfiles==1.0.4
websockets==14.2
